==> notebooks/05_visualize_training_data.ipynb <==


import pandas as pd

from paths import TRANSFORMED_DATA_DIR

features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR/'tabular_data.parquet')
features_and_targets


features = features_and_targets.drop(columns = ['target_rides_next_hour'], axis =1)
targets =  features_and_targets['target_rides_next_hour']

features.shape, targets.shape

%reload_ext autoreload
%autoreload 2

import sys
sys.path.insert(0, '/Users/rajesh/taxi_demand_predictor_1/src')

print(sys.path)

from plot import plot_one_sample

plot_one_sample(
    features, 
    targets,
    example_id = 5
)
