==> notebooks/06_baseline_model.ipynb <==
import sys
sys.path.insert(0, '../src')

import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR/'tabular_data.parquet')

df

from datetime import datetime
from data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df, 
    cutoff_date = datetime(2022, 7,1,0,0,0),
    target_column_name = 'target_rides_next_hour'
)

X_train.shape, y_train.shape, X_test.shape, y_test.shape

import numpy as np
class BaselineModelPreviousHour:
    """
    Prediction = actual demand observered in last hour
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame):
        return X_test[f'rides_previous_1_hour']


model = BaselineModelPreviousHour()
predictions = model.predict(X_test)
predictions

%reload_ext autoreload
%autoreload 2

pip install scikit-learn

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test,predictions)
print(f'{test_mae =:.4f}')

import pandas as pd
import numpy as np

class BaselineModelPreviousWeek:
    """
    Prediction = actual demand observed at t-7 days
    """

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_train: pd.DataFrame):
        return X_test[f'rides_previous_{7*24}_hour']
        

model1 = BaselineModelPreviousWeek()

prediction1= model1.predict(X_test)
test_mae1 = mean_absolute_error(y_test, prediction1)

print(f'{test_mae1 =:.4f}')



class BaselineModelLast4Weeks:
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame):
        return 0.25*(
            X_test[f'rides_previous_{7*24}_hour']+ \
            X_test[f'rides_previous_{2*7*24}_hour']+ \
            X_test[f'rides_previous_{3*7*24}_hour']+ \
            X_test[f'rides_previous_{4*7*24}_hour']
        )

model2 = BaselineModelLast4Weeks()

prediction2= model2.predict(X_test)
test_mae2 = mean_absolute_error(y_test, prediction2)

print(f'{test_mae2 =:.4f}')
