==> notebooks/08_lightGBM_model.ipynb <==
import sys
sys.path.insert(0, '../src')

import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR/'tabular_data.parquet')

df

from datetime import datetime
from data_split import train_test_split

X_train, y_train, X_test, y_test =  train_test_split(
    df, 
    cutoff_date = datetime(2022, 7, 1, 0,0,0),
    target_column_name = 'target_rides_next_hour'
)

X_train.shape, y_train.shape, X_test.shape, y_test.shape

#use only past data and remove hour data
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]
X_train_only_numeric.shape

pip uninstall lightgbm


pip install --no-binary :all: lightgbm


import lightgbm as lgb

model = lgb.LGBMRegressor()

model.fit(X_train_only_numeric, y_train)


from sklearn.metrics import mean_absolute_error

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
mae_xgb = mean_absolute_error(y_test, predictions)

print(f'{mae_xgb=:.4f}')
