==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
import sys
sys.path.insert(0, '../src')

import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR/'tabular_data.parquet')

df

from datetime import datetime
from data_split import train_test_split

X_train, y_train, X_test, y_test =  train_test_split(
    df, 
    cutoff_date = datetime(2022, 7, 1, 0,0,0),
    target_column_name = 'target_rides_next_hour'
)

X_train.shape, y_train.shape, X_test.shape, y_test.shape

def average_rides_last_4_weeks(X: pd.DataFrame)->pd.DataFrame:
    """
    Adds the column with average rides for 7, 14, 21 and 28 days previous
    """
    X['average_rides_last_4_weeks'] = 0.25*(
            X[f'rides_previous_{7*24}_hour']+ \
            X[f'rides_previous_{2*7*24}_hour']+ \
            X[f'rides_previous_{3*7*24}_hour']+ \
            X[f'rides_previous_{4*7*24}_hour']
    )

    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate = False
)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

X_train

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatureEngineer(BaseEstimator, TransformerMixin):
    def fit(self, X, y = None):
        return self

    def transform(self, X):
        X_ = X.copy()

        X_['hour'] = X_['pickup_hour'].dt.hour
        X_['day_of_week'] = X_['pickup_hour'].dt.dayofweek

        return X_.drop(columns = ['pickup_hour'])

add__temporal_features = TemporalFeatureEngineer()
add__temporal_features.fit_transform(X_train)

import lightgbm as lgb
from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add__temporal_features,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

from sklearn.metrics import mean_absolute_error

predictions = pipeline.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)

print(f'{test_mae=:.4f}')


# pip install lightgbm

import lightgbm as lgb
